some initial changes to remove tmp service w/ burnettk
