+++ spiffworkflow-backend/src/spiffworkflow_backend/services/script_unit_test_runner.py
@@ -7,7 +7,7 @@ from typing import Any
 from SpiffWorkflow.bpmn.exceptions import WorkflowTaskException  # type: ignore
 from SpiffWorkflow.task import Task as SpiffTask  # type: ignore
 
-from spiffworkflow_backend.services.process_instance_processor import CustomBpmnScriptEngine
+from spiffworkflow_backend.services.process_instance_script_engine import CustomBpmnScriptEngine
 
 PythonScriptContext = dict[str, Any]
